backend/requirements.txt
numpy
scipy
pyaudio
Flask
flask_cors
pyfluidsynth
music21